+++ src/arguments/index.ts
@@ -2,7 +2,9 @@ import { dofusCoordinates } from './dofus/coordinates';
 import { dofusDimension } from './dofus/dimension';
 import { dofusJob } from './dofus/job';
 import { dofusServer } from './dofus/server';
+import { date } from './time/date';
 import { duration } from './time/duration';
+import { time } from './time/time';
 import { line } from './text/line';
 import { oneline } from './text/oneline';
 
@@ -11,7 +13,9 @@ export const argumentTypes = {
   dofusDimension,
   dofusJob,
   dofusServer,
+  date,
   duration,
+  time,
   line,
   oneline
 };